Move civ init functions
